Only attempt to move files if source exists instead of throwing exception

+++ node-admin-maintenance/src/main/java/com/yahoo/vespa/hosted/node/maintenance/Maintainer.java
@@ -64,7 +64,9 @@ public class Maintainer {
             case "move-files":
                 Path from = Paths.get((String) job.getRequiredArgument("from"));
                 Path to = Paths.get((String) job.getRequiredArgument("to"));
-                Files.move(from, to);
+                if (Files.exists(from)) {
+                    Files.move(from, to);
+                }
                 break;
 
             case "handle-core-dumps":
